requirements.txt
Flask==3.1.1
flask-jwt-extended==4.7.1
python-dotenv==1.1.1
PyMySQL==1.1.1
pymongo==4.16.0
bcrypt==4.3.0
google-genai==1.64.0